Improved the Player Movement
- Refined the player's movement logic.
- Organized the project with filters for the various classes.

+++ Graficos1_TP3/Bullet.h
@@ -6,7 +6,7 @@
 class Bullet : public Entity
 {
 private:
-	float _speed;
+	int _speed;
 	bool _enabled;
 	
 public:
